+++ library/includes/classes/class-admin.php
@@ -6,7 +6,7 @@
  * @subpackage  Admin
  *
  * @since    1.0
- * @version  1.8.1
+ * @version  1.9
  *
  * Contents:
  *
@@ -87,7 +87,7 @@ final class {%= prefix_class %}_Theme_Framework_Admin {
 		 * Admin assets
 		 *
 		 * @since    1.0
-		 * @version  1.5
+		 * @version  1.9
 		 */
 		public static function assets() {
 
@@ -98,7 +98,7 @@ final class {%= prefix_class %}_Theme_Framework_Admin {
 					// Styles
 
 						$register_styles = apply_filters( 'wmhook_{%= prefix_hook %}_tf_admin_assets_register_styles', array(
-								'{%= prefix_var %}-welcome' => array( {%= prefix_class %}_Theme_Framework::get_stylesheet_directory_uri( {%= prefix_constant %}_LIBRARY_DIR . 'css/welcome.css' ) ),
+								'{%= prefix_var %}-welcome' => array( {%= prefix_class %}_Theme_Framework::get_stylesheet_directory_uri( 'library/css/welcome.css' ) ),
 							) );
 
 						foreach ( $register_styles as $handle => $atts ) {
@@ -145,7 +145,7 @@ final class {%= prefix_class %}_Theme_Framework_Admin {
 		 *   );
 		 *
 		 * @since    1.0
-		 * @version  1.1
+		 * @version  1.9
 		 */
 		public static function message() {
 
@@ -172,7 +172,7 @@ final class {%= prefix_class %}_Theme_Framework_Admin {
 
 				$class      = 'updated';
 				$repeat     = 0;
-				$capability = apply_filters( 'wmhook_{%= prefix_hook %}_tf_admin_message_capability', 'switch_themes' );
+				$capability = apply_filters( 'wmhook_{%= prefix_hook %}_tf_admin_message_capability', 'edit_theme_options' );
 				$message    = get_transient( '{%= prefix_var %}_admin_notice' );