+++ library/includes/classes/class-customize-control-html.php
@@ -8,9 +8,9 @@
  * @subpackage  Customize
  *
  * @since    1.0
- * @version  1.8
+ * @version  1.9
  */
-class {%= prefix_class %}_Control_HTML extends WP_Customize_Control {
+class {%= prefix_class %}_Customize_Control_HTML extends WP_Customize_Control {
 
 	public $type = 'html';
 
@@ -38,4 +38,4 @@ class {%= prefix_class %}_Control_HTML extends WP_Customize_Control {
 
 	} // /render_content
 
-} // /{%= prefix_class %}_Control_HTML
+} // /{%= prefix_class %}_Customize_Control_HTML