+++ library/includes/classes/class-customize-control-multiselect.php
@@ -8,9 +8,9 @@
  * @subpackage  Customize
  *
  * @since    1.0
- * @version  1.8
+ * @version  1.9
  */
-class {%= prefix_class %}_Control_Multiselect extends WP_Customize_Control {
+class {%= prefix_class %}_Customize_Control_Multiselect extends WP_Customize_Control {
 
 	public $type = 'multiselect';
 
@@ -49,4 +49,4 @@ class {%= prefix_class %}_Control_Multiselect extends WP_Customize_Control {
 
 	} // /render_content
 
-} // /{%= prefix_class %}_Control_Multiselect
+} // /{%= prefix_class %}_Customize_Control_Multiselect