+++ library/includes/classes/class-visual-editor.php
@@ -9,7 +9,7 @@
  * @subpackage  Visual Editor
  *
  * @since    1.0
- * @version  1.8.1
+ * @version  1.9
  *
  * Contents:
  *
@@ -444,7 +444,7 @@ final class {%= prefix_class %}_Theme_Framework_Visual_Editor {
 		 * Adding scripts to post edit screen
 		 *
 		 * @since    1.7.2
-		 * @version  1.8
+		 * @version  1.9
 		 *
 		 * @param  string $hook_suffix
 		 */
@@ -465,7 +465,7 @@ final class {%= prefix_class %}_Theme_Framework_Visual_Editor {
 
 					wp_enqueue_script(
 							'{%= prefix_var %}-post-edit',
-							{%= prefix_class %}_Theme_Framework::get_stylesheet_directory_uri( {%= prefix_constant %}_LIBRARY_DIR . 'js/post.js' ),
+							{%= prefix_class %}_Theme_Framework::get_stylesheet_directory_uri( 'library/js/post.js' ),
 							array( 'jquery' ),
 							esc_attr( {%= prefix_constant %}_THEME_VERSION ),
 							true